+++ pyccel/ast/class_defs.py
@@ -15,7 +15,7 @@ from .builtins   import PythonImag, PythonReal, PythonConjugate
 from .core       import ClassDef, PyccelFunctionDef
 from .datatypes  import (PythonNativeBool, PythonNativeInt, PythonNativeFloat,
                          PythonNativeComplex, StringType, TupleType, CustomDataType,
-                         HomogeneousListType, HomogeneousSetType)
+                         HomogeneousListType, HomogeneousSetType, SymbolicType)
 from .numpyext   import (NumpyShape, NumpySum, NumpyAmin, NumpyAmax,
                          NumpyImag, NumpyReal, NumpyTranspose,
                          NumpyConjugate, NumpySize, NumpyResultType, NumpyArray)
@@ -254,6 +254,8 @@ def get_cls_base(class_type):
         return ListClass
     elif isinstance(class_type, HomogeneousSetType):
         return SetClass
+    elif isinstance(class_type, SymbolicType):
+        return None
     else:
         raise NotImplementedError(f"No class definition found for type {class_type}")