+++ tests/errors/semantic/non_blocking/wrong_class_type_annotation.py
@@ -11,4 +11,3 @@ class A:
 if __name__ == '__main__':
     myA : int = A()
 
-    print(myA.x)